fix: Move Page to components
